Implement a cross-platform numbering FilterWriter
Character-wise operations are required to perform line numbering. In
particular, it is necessary to keep information about the previous
characters, to make sure the sequence '\r\n' only creates a single new
line, and not two different lines.

I've chosen to implement this with a small state machine, with four
different states :

+ Empty.
+ Last char is '\r'.
+ Last char is '\n'.
+ Last char is a standard char.

When a '\n' is detected, output is performed aggressively. Otherwise, I
wait until some additional char(s) is entered to make a choice as to
whether the line number should be written or not.